+++ SmileBoy.Client.BL/Managers/ProductManager.cs
@@ -11,11 +11,13 @@ namespace SmileBoy.Client.Core.Managers
 {
     public class ProductManager : IProductManager
     {
+        private readonly IReferenceExcludable _excludable;
         private readonly IProductService _productService;
         private readonly IUnitOfWork _unitOfWork;
 
-        public ProductManager(IUnitOfWork unitOfWork, IProductService productService)
+        public ProductManager(IUnitOfWork unitOfWork, IReferenceExcludable excludable, IProductService productService)
         {
+            _excludable = Has.NotNull(excludable);
             _productService = Has.NotNull(productService);
             _unitOfWork = Has.NotNull(unitOfWork);
         }
@@ -26,16 +28,17 @@ namespace SmileBoy.Client.Core.Managers
         /// <returns></returns>
         public async Task DeleteAsync(Guid id)
         {
-            //soon
+            var entity = await _unitOfWork.Products.GetByIdAsync(id);
 
-            //var entity = await _unitOfWork.Products.GetByIdAsync(id);
-            //Has.NotNull(entity);
+            if (entity == null)
+                throw new InvalidOperationException();
 
+            await _excludable.Exclude(o => o.DeleteReferencesProduct(id), entity);
             await _productService.DeleteAsync(id); 
         }
 
         public async Task<PaginationResult<ProductDto>> GetAll(int page, int pageSize, string filter = null) =>
-            await _productService. GetAll(page, pageSize, filter);
+            await _productService.GetAll(page, pageSize, filter);
 
         public async Task<ProductDto> GetByIdAsync(Guid id) =>
             await _productService.GetByIdAsync(id);
